% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swell.r
\name{swell}
\alias{swell}
\title{Generate and optimize a braille-embedded ready-to-print pdf file for swell form machine}
\usage{
swell(x, file = "output.pdf", lines = c(1))
}
\arguments{
\item{x}{the call to create a graph}

\item{file}{A character string giving the filename where the image is to be saved. The default file name is "output.pdf" if not given.}

\item{lines}{a vector argument that is passed to lty. Specify this as c(1, 3) when two lines are necessary (e.g., ANOVA interaction plot); and c(1, 3, 5) for three lines required plots.}
}
\value{
Nothing within the R session, but a pdf file will be created in the user's working directory.
}
\description{
The first argument to this function must be a call to create a graph, such as a histogram. Instead of opening a new graphics device, the graph will be created in a pdf file, with all text being presented using  a braille font.
}
\details{
swell


*Note*: This function optimizes for both 11*11 and 11*8.5 swell touch papers. When printing a generated braille pdf, just remember two things:
1. check "Fit" in the print dialog box (the shortcut key is ALT+F).
2. Check "Landscape" in the print dialog box as well.
* For boxplot, you do not need to check the "Landscape," just leave it as a "Portray." However, please check the "Fit" all the time.
*Information*:
If you are not familiar with what swell form machine is, please refer to [this web site](http://www.americanthermoform.com/product/swell-form-graphics-ii-machine/).
}
\examples{
swell(hist(airquality$Ozone), "test.pdf")
# The braille ready-to-print pdf version will be created in your working directory. Print it out and run the swell touch paper through swell form heating machine.
}
\author{
JooYoung Seo
}
\keyword{swell}
